infra/requirements.txt
aws-cdk-lib>=2.240.0
constructs>=10.5.1
jsii>=1.126.0
